Service deployment

+++ node-app-cicd.sh
@@ -15,7 +15,9 @@
 # $5 npm/node ci/install command (defaults to "npm ci")
 # $6 npm/node test command (defaults to "npm test", optional)
 # $7 npm/node bundle/debundle command (defaults to "", optional)
-# $8 Temporary directory for deployment backup and other temp files (defaults to "/tmp", DEPLOY only)
+# $8 node environment that will be used to set NODE_ENV (defaults to "test", DEPLOY only)
+# $9 node app starting port for service (increments int to number of physical cores, DEPLOY only)
+# $10 Temporary directory for deployment backup and other temp files (defaults to "/tmp", DEPLOY only)
 
 EXEC_TYPE=`[[ ("$1" == "BUILD" || "$1" == "DEPLOY") ]] && echo $1 || echo ""`
 APP_NAME=`[[ (-n "$2") ]] && echo $2`
@@ -24,7 +26,9 @@ NVMRC_DIR=`[[ (-n "$4") ]] && echo $4 || echo "/opt"`
 CMD_INSTALL=`[[ (-n "$5") ]] && echo $5 || echo "npm ci"`
 CMD_TEST=`[[ (-n "$6") ]] && echo $6 || echo "npm test"`
 CMD_BUNDLE=`[[ (-n "$7") ]] && echo $7 || echo ""`
-APP_TMP=`[[ (-n "$8") ]] && echo $8 || echo /tmp`
+APP_PORT_START=`[[ "$8" =~ ^[0-9]+$ ]] && echo $8 || echo ""`
+NODE_ENV=`[[ (-n "$9") ]] && echo $9 || echo "test"`
+APP_TMP=`[[ (-n "${10}") ]] && echo ${10} || echo /tmp`
 
 execCmdCICD () {
   if [[ (-n "$1") ]]; then
@@ -46,18 +50,18 @@ else
   echo "Missing or invalid execution type (first argument, either \"BUILD\" or \"DEPLOY\"" >&2
   exit 1
 fi
-if [[ (-n "$APP_NAME") ]]; then
-  echo "$EXEC_TYPE: using app name $APP_NAME"
-else
-  echo "$EXEC_TYPE: missing app name argument" >&2
+if [[ "$APP_NAME" =~ [^a-zA-Z] ]]; then
+  echo "$EXEC_TYPE: missing or invalid app name \"$APP_NAME\" (must contain only alpha characters)" >&2
   exit 1
+else
+  echo "$EXEC_TYPE: using app name $APP_NAME"
 fi
 if [[ (-d "$APP_DIR") ]]; then
   echo "$EXEC_TYPE: using app dir $APP_DIR"
   if [[ ("$EXEC_TYPE" == "DEPLOY") ]]; then
     echo "$EXEC_TYPE: backing up $APP_DIR ..."
     tar -czf $APP_TMP/$APP_NAME-backup-`date +%Y%m%d_%H%M%S`.tar.gz $APP_DIR/*
-    rm -rf $APP_DIR/*
+    [[ $? != 0 ]] && { echo "$EXEC_TYPE: failed to backup $APP_DIR to $APP_TMP" >&2; exit 1; }
   fi
 elif [[ ("$EXEC_TYPE" == "BUILD") ]]; then
   echo "$EXEC_TYPE: unable to find dir $APP_DIR" >&2
@@ -68,13 +72,38 @@ elif [[ (-z "$APP_DIR") ]]; then
 else
   # DEPLOY: create new app dir
   sudo mkdir -p $APP_DIR
-  sudo chown -hR $USER $APP_DIR
 fi
 if [[ ("$EXEC_TYPE" == "DEPLOY") ]]; then
+  # check if the service is installed
+  if [[ -n "$APP_PORT" ]]; then
+    TARGETED=`sudo systemctl list-units --all -t target --full --no-legend | grep "$APP_NAME.target"`
+    if [[ -z "$TARGETED" ]]; then
+      # match the number of processes/services with the number of physical cores
+      CORE_CNT=`getconf _NPROCESSORS_ONLN`
+      for (( c=$APP_PORT; c<=$CORE_CNT + $APP_PORT; c++ )); do
+        echo "$EXEC_TYPE: checking if port $c is in use"
+        PORT_USED=`sudo ss -tulwnH "( sport = :$c )"`
+        if [[ -n "$PORT_USED" ]]; then
+          echo "$EXEC_TYPE: app port $c is already in use (core count: $CORE_CNT, start port: $APP_PORT)" >&2
+          exit 1
+        fi
+        SERVICES=`[[ -n "$SERVICES" ]] && echo "$SERVICES " || echo ""`
+        SERVICES="$SERVICES$APP_NAME@$c.service"
+      done
+    else
+      sudo systemctl stop $APP_NAME.target
+      [[ $? != 0 ]] && { echo "$EXEC_TYPE: failed to stop $APP_NAME.target" >&2; exit 1; }
+    fi
+  fi
   sudo chown -hR $USER $APP_DIR
+  # replace app contents with extracted content
   if [[ (-f "$APP_TMP/$APP_NAME.tar.gz") ]]; then
-    echo "$EXEC_TYPE: extracting app contents from \"$APP_TMP/$APP_NAME.tar.gz\" to \"$APP_DIR\""
-    tar --warning=no-timestamp -xzvf $APP_TMP/$APP_NAME.tar.gz -C $APP_DIR
+    echo "$EXEC_TYPE: cleaning app at $APP_DIR"
+    sudo rm -rfd $APP_DIR/*
+    [[ $? != 0 ]] && { echo "$EXEC_TYPE: failed to clean $APP_DIR" >&2; exit 1; }
+    echo "$EXEC_TYPE: extracting app contents from $APP_TMP/$APP_NAME.tar.gz to $APP_DIR"
+    tar --warning=no-timestamp --strip-components=1 -xzvf $APP_TMP/$APP_NAME.tar.gz -C $APP_DIR
+    [[ $? != 0 ]] && { echo "$EXEC_TYPE: failed to extract $APP_TMP/$APP_NAME.tar.gz to $APP_DIR" >&2; exit 1; }
     # remove extracted app archive
     sudo rm -f $APP_TMP/$APP_NAME.tar.gz
   else
@@ -103,6 +132,57 @@ elif [[ (-z "$NVMRC_VER") ]]; then
   exit 1
 fi
 
+# DEPLOY: service/target templates
+# ExecStart=/bin/bash -c '$NVM_DIR/nvm_exec node .'
+SERVICE=`[[ -z "$SERVICES" ]] && echo "" || echo "
+# /etc/systemd/system/$APP_NAME@.service
+[Unit]
+Description=\"$APP_NAME (%H:%i)\"
+After=network.target
+# Wants=redis.service
+PartOf=$APP_NAME.target
+
+[Service]
+Environment=NODE_ENV=$NODE_ENV
+Environment=NODE_HOST=%H
+Environment=NODE_PORT=%i
+Type=simple
+# user should match the user where nvm was installed
+User=$USER
+WorkingDirectory=$APP_DIR
+# run node using the node version defined in working dir .nvmrc
+ExecStart=/bin/bash -c '~/.nvm/nvm_exec node .'
+Restart=on-failure
+RestartSec=5
+StandardError=syslog
+
+[Install]
+WantedBy=multi-user.target
+"`
+TARGET=`[[ -z "$SERVICES" ]] && echo "" || echo "
+# /etc/systemd/system/$APP_NAME.target
+[Unit]
+Description=\"$APP_NAME\"
+Wants=$SERVICES
+
+[Install]
+WantedBy=multi-user.target
+"`
+
+if [[ -n "$SERVICE" && -n "$TARGET" ]]; then
+  SERVICE_PATH=`/etc/systemd/system/$APP_NAME@.service`
+  TARGET_PATH=`/etc/systemd/system/$APP_NAME.target`
+  sudo echo "$EXEC_TYPE: creating $SERVICE_PATH and $TARGET_PATH"
+  sudo echo "$SERVICE" > "$SERVICE_PATH"
+  [[ $? != 0 ]] && { echo "$EXEC_TYPE: failed to write $SERVICE_PATH" >&2; exit 1; }
+  sudo echo "$TARGET" > "$TARGET_PATH"
+  [[ $? != 0 ]] && { echo "$EXEC_TYPE: failed to write $TARGET_PATH" >&2; exit 1; }
+  sudo systemctl daemon-reload
+  [[ $? != 0 ]] && { echo "$EXEC_TYPE: failed to reload the service/target daemon" >&2; exit 1; }
+  sudo systemctl enable $APP_NAME.target
+  [[ $? != 0 ]] && { echo "$EXEC_TYPE: failed to enable $TARGET_PATH" >&2; exit 1; }
+fi
+
 # enable nvm (alt "$NVM_DIR/nvm-exec node" or "$NVM_DIR/nvm-exec npm")
 #NVM_EDIR=`[[ (-n "$NVM_DIR") ]] && echo $NVM_DIR || echo "$HOME/.nvm"`
 #if [[ (-x "$NVM_EDIR/nvm-exec") ]]; then
@@ -129,6 +209,7 @@ if [[ ("$EXEC_TYPE" == "BUILD") ]]; then
   [[ $? != 0 ]] && exit 1
   # create app archive
   tar --exclude='./*git*' --exclude='./node_modules' --exclude='*.gz' -czvf $APP_NAME.tar.gz .
+  [[ $? != 0 ]] && { echo "$EXEC_TYPE: failed to create app archive $APP_NAME.tar.gz" >&2; exit 1; }
 else
   # execute debundle
   execCmdCICD "$CMD_BUNDLE" "debundling"
@@ -136,6 +217,11 @@ else
   # execute install
   execCmdCICD "$CMD_INSTALL" "ci/install"
   [[ $? != 0 ]] && exit 1
+  # start the services
+  sudo systemctl start $APP_NAME.target
+  [[ $? != 0 ]] && { echo "$EXEC_TYPE: failed to start $APP_NAME.target" >&2; exit 1; }
+  SERVICE_STARTED=`sudo systemctl is-active "$APP_NAME" >/dev/null 2>&1 && echo ACTIVE || echo ""`
+  [[ -z "$SERVICE_STARTED" ]] && { echo "$EXEC_TYPE: $APP_NAME.target is not active" >&2; exit 1; }
   # execute tests
   execCmdCICD "$CMD_TEST" "tests"
   [[ $? != 0 ]] && exit 1
